+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [1.1.2]
+
+### Added
+- initial warning setup confirmation.
+
+### Changed
+- Updated CLI commands for new setup.
+
 ## [1.1.1]
 
 ### Added
@@ -26,6 +34,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 Initial tagged release.
 
+[1.1.2]: https://github.com/infinum/eightshift-create/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/infinum/eightshift-create/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/infinum/eightshift-create/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/infinum/eightshift-create/releases/tag/1.0.0